refactor: Member 도메인 역할 분리
