+++ src/main/java/cart/application/MemberService.java
@@ -1,10 +1,11 @@
 package cart.application;
 
 import cart.db.repository.MemberRepository;
-import cart.domain.member.Member;
+import cart.domain.member.*;
 import cart.dto.login.MemberRequest;
 import cart.dto.login.TokenResponse;
 import cart.exception.AuthenticationException;
+import cart.util.BasicTokenGenerator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,16 +31,19 @@ public class MemberService {
         if (memberRepository.existsByName(memberRequest.getName())) {
             throw new AuthenticationException(DUPLICATED_NAME);
         }
-        Long memberId = memberRepository.save(new Member(memberRequest.getName(), memberRequest.getPassword()));
+
+        Password encryptedPassword = Password.encrypt(memberRequest.getPassword());
+        Long memberId = memberRepository.save(new Member(memberRequest.getName(), encryptedPassword));
         Member newMember = new Member(memberId, memberRequest.getName(), memberRequest.getPassword());
         memberCouponService.add(newMember, CONGRATULATION_COUPON_ID_TO_NEW_MEMBER);
     }
 
     public TokenResponse generateMemberToken(final MemberRequest memberRequest) {
-        Member member = new Member(memberRequest.getName(), memberRequest.getPassword());
+        Password encryptedPassword = Password.encrypt(memberRequest.getPassword());
+        Member member = new Member(memberRequest.getName(), encryptedPassword);
         if (!memberRepository.existsByMember(member)) {
             throw new AuthenticationException(NOT_AUTHENTICATION_MEMBER);
         }
-        return new TokenResponse(member.generateToken());
+        return new TokenResponse(BasicTokenGenerator.generate(member.getName(), member.getPassword()));
     }
 }
